Replaced burl-based auth with proper header-based auth
Improved logging (no more println)
added integration tests for auth, need some proper unit tests

+++ integration_tests.sh
@@ -6,8 +6,16 @@
 set -euo pipefail
 
 function exit_script {
-  echo "Cleaning up"
-  kill "$bibin_pid"
+  if [[ ${bibin_pid-} != "" ]]; then
+    echo "Cleaning up $bibin_pid"
+    kill "$bibin_pid"
+  else
+    echo "No serevr was started"
+  fi
+}
+
+function error_script {
+  echo "ERROR on line ${BASH_LINENO[0]}"
 }
 
 function assert_equal {
@@ -17,13 +25,14 @@ function assert_equal {
     fi
 }
 
+trap error_script ERR
 trap exit_script EXIT
 
 prefix=http://localhost:8000/
 password=a69711b1-3d39-4344-b97a-ba91e2f5adca
 browser_agent="Mozilla/5.0 (X11; Linux x86_64; rv:76.0) Gecko/20100101 Firefox/76.0"
 
-env "ROCKET_PASSWORD=$password" "ROCKET_PREFIX=$prefix" ROCKET_ENV=production cargo run --release&
+env "ROCKET_PASSWORD=$password" ROCKET_LOG="normal" "ROCKET_PREFIX=$prefix" ROCKET_ENV=production cargo run --release&
 bibin_pid=$!
 
 while ! curl -s "$prefix" > /dev/null; do
@@ -31,10 +40,17 @@ while ! curl -s "$prefix" > /dev/null; do
 done
 echo "Bibin started, testing"
 
-echo "#### Check that after uploading some data with curl, we can get it back as text with curl"
+echo "#### Check that after uploading some data with curl (X-API-Key header), we can get it back as text with curl"
 sample_data1="hello world"
-url="$(curl -X PUT --data "$sample_data1" "$prefix$password")"
+url="$(curl -X PUT -H "X-API-Key: $password" --data "$sample_data1" "$prefix")"
 assert_equal "$(curl -s "$url")" "$sample_data1"
+
+echo "#### Check that after uploading some data with curl (Authorization header), we can get it back as text with curl"
+sample_data1="hello world"
+url="$(curl -X PUT -u "a:$password" --data "$sample_data1" "$prefix")"
+assert_equal "$(curl -s "$url")" "$sample_data1"
+
+
 
 echo "#### Check that after uploading some data with curl, we can get it back as html with a browser"
 assert_equal "$(curl -H "User-Agent: $browser_agent" -s "$url" | head -n1)"  "<!DOCTYPE html>"
@@ -54,7 +70,7 @@ assert_equal "$data_after_redirect" "$sample_data2"
 
 echo "#### Check that it can handle 10'000 requests (There will be ID collisions)"
 num=0
-while test "$num" -lt 2500; do
+while test "$num" -lt 250; do
     curl -L -s -d password="$password" -d val="$sample_data1 - $num" "$prefix" > /dev/null&
     bg_process1="$!"
 
@@ -64,10 +80,17 @@ while test "$num" -lt 2500; do
     curl -L -s -d password="$password" -d val="$sample_data1 - $num" "$prefix" > /dev/null&
     bg_process3="$!"
 
-    url="$(curl -s -X PUT --data "$sample_data2 - $num" "$prefix$password")"
+    url="$(curl -s -X PUT -H "X-API-Key: $password" --data "$sample_data2 - $num" "$prefix")"
     assert_equal "$(curl -s "$url")" "$sample_data2 - $num"
     wait "$bg_process1" "$bg_process2" "$bg_process3"
     num="$(( num + 1 ))"
 done
+
+echo "#### Testing invalid credentials"
+if curl -s -f -X PUT -H "X-API-Key: dummy" --data "$sample_data2 - $num" "$prefix"; then false; fi
+if curl -s -f -X PUT -H "X-API-Key" --data "$sample_data2 - $num" "$prefix"; then false; fi
+if curl -s -f -X PUT -H "Authorization: basic ffffffff" --data "$sample_data2 - $num" "$prefix"; then false; fi
+if curl -s -f -X PUT -u "b:" --data "$sample_data2 - $num" "$prefix"; then false; fi
+if curl -s -f -X PUT -H "X-API-Key: $password\0" --data "$sample_data2 - $num" "$prefix"; then false; fi
 
 echo "All tests OK"
